// codigo/frontend/valemax/src/enums/node-type.enum.ts
/**
 * Represents the node types of the graph.
 */
export enum NodeType {
    Cliente = 'cliente',
    Patio = 'patio',
    Porto = 'porto',
    SiteProdutivo = 'siteProdutivo',
    UsinaBeneficiamento = 'usinaBeneficiamento',
    UsinaBriquete = 'usinaBriquete',
    UsinaPelotizacao = 'usinaPelotizacao',
  }